Implement hook to verify user's authentication status
